Backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt>=4.3.0",
    "fastapi>=0.116.1",
    "groq>=0.30.0",
    "httpx>=0.23.0",
    "nltk>=3.9.1",
    "numpy>=2.3.1",
    "passlib[bcrypt]>=1.7.4",
    "phonenumbers>=9.0.11",
    "pydantic[email]>=2.11.7",
    "pypdf2>=3.0.1",
    "python-docx>=1.2.0",
    "python-dotenv>=1.1.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "scikit-learn>=1.7.0",
    "sqlalchemy>=2.0.41",
    "supabase>=2.18.0",
    "uvicorn>=0.35.0",
]

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "httpx>=0.23.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*", "schemas*"]
